Updated to commit 15290a0613

+++ src/app/code/core/Mage/Admin/Model/Resource/User.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
  * OpenMage
  *
@@ -10,7 +9,7 @@
  * @category   Mage
  * @package    Mage_Admin
  * @copyright  Copyright (c) 2006-2020 Magento, Inc. (https://www.magento.com)
- * @copyright  Copyright (c) 2017-2024 The OpenMage Contributors (https://www.openmage.org)
+ * @copyright  Copyright (c) 2017-2023 The OpenMage Contributors (https://www.openmage.org)
  * @license    https://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
 
@@ -35,11 +34,11 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
         $this->_uniqueFields = [
             [
                 'field' => 'email',
-                'title' => Mage::helper('adminhtml')->__('Email'),
+                'title' => Mage::helper('adminhtml')->__('Email')
             ],
             [
                 'field' => 'username',
-                'title' => Mage::helper('adminhtml')->__('User Name'),
+                'title' => Mage::helper('adminhtml')->__('User Name')
             ],
         ];
         return $this;
@@ -48,6 +47,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Authenticate user by $username and $password
      *
+     * @param Mage_Admin_Model_User $user
      * @return $this
      */
     public function recordLogin(Mage_Admin_Model_User $user)
@@ -56,7 +56,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
 
         $data = [
             'logdate' => Varien_Date::now(),
-            'lognum'  => $user->getLognum() + 1,
+            'lognum'  => $user->getLognum() + 1
         ];
 
         $condition = [
@@ -83,7 +83,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             ->where('username=:username');
 
         $binds = [
-            'username' => $username,
+            'username' => $username
         ];
 
         return $adapter->fetchRow($select, $binds);
@@ -135,6 +135,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
         if ($user->isObjectNew()) {
             $user->setCreated($this->formatDate(true));
         }
+
         $user->setModified($this->formatDate(true));
 
         return parent::_beforeSave($user);
@@ -143,6 +144,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Unserialize user extra data after user save
      *
+     * @param Mage_Core_Model_Abstract $user
      * @return $this
      */
     protected function _afterSave(Mage_Core_Model_Abstract $user)
@@ -164,6 +166,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Delete user role record with user
      *
+     * @param Mage_Core_Model_Abstract $user
      * @return $this
      * @throws Exception
      */
@@ -176,7 +179,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
         $adapter->beginTransaction();
         try {
             $conditions = [
-                'user_id = ?' => $uid,
+                'user_id = ?' => $uid
             ];
 
             $adapter->delete($this->getMainTable(), $conditions);
@@ -186,6 +189,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             $adapter->rollBack();
             throw $e;
         }
+
         $this->_afterDelete($user);
         return $this;
     }
@@ -193,6 +197,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * TODO: unify _saveRelations() and add() methods, they make same things
      *
+     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
      * @return $this|Mage_Core_Model_Abstract
      */
     public function _saveRelations(Mage_Core_Model_Abstract $user)
@@ -213,11 +218,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             $adapter->delete($this->getTable('admin/role'), $conditions);
             foreach ($rolesIds as $rid) {
                 $rid = (int) $rid;
-                if ($rid > 0) {
-                    $role = Mage::getModel('admin/role')->load($rid);
-                } else {
-                    $role = new Varien_Object(['tree_level' => 0]);
-                }
+                $role = $rid > 0 ? Mage::getModel('admin/role')->load($rid) : new Varien_Object(['tree_level' => 0]);
 
                 $data = new Varien_Object([
                     'parent_id'  => $rid,
@@ -225,7 +226,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
                     'sort_order' => 0,
                     'role_type'  => Mage_Admin_Model_Acl::ROLE_TYPE_USER,
                     'user_id'    => $user->getId(),
-                    'role_name'  => $user->getFirstname(),
+                    'role_name'  => $user->getFirstname()
                 ]);
 
                 $insertData = $this->_prepareDataForTable($data, $this->getTable('admin/role'));
@@ -236,11 +237,9 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
                 // reload acl on next user http request
                 $this->saveReloadAclFlag($user, 1);
             }
+
             $adapter->commit();
-        } catch (Mage_Core_Exception $e) {
-            $adapter->rollBack();
-            throw $e;
-        } catch (Exception $e) {
+        } catch (Mage_Core_Exception|Exception $e) {
             $adapter->rollBack();
             throw $e;
         }
@@ -251,6 +250,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Get user roles
      *
+     * @param Mage_Core_Model_Abstract $user
      * @return array
      */
     public function getRoles(Mage_Core_Model_Abstract $user)
@@ -266,7 +266,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
                     ->joinLeft(
                         ['ar' => $table],
                         "(ar.role_id = {$table}.parent_id and ar.role_type = 'G')",
-                        ['role_id'],
+                        ['role_id']
                     )
                     ->where("{$table}.user_id = :user_id");
 
@@ -286,17 +286,18 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Save user roles
      *
+     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
      * @return $this
      */
     public function add(Mage_Core_Model_Abstract $user)
     {
         $dbh = $this->_getWriteAdapter();
         $aRoles = $this->hasAssigned2Role($user);
-        if (count($aRoles)) {
-            foreach ($aRoles as $idx => $data) {
+        if (is_array($aRoles) && count($aRoles)) {
+            foreach ($aRoles as $data) {
                 $dbh->delete(
                     $this->getTable('admin/role'),
-                    ['role_id = ?' => $data['role_id']],
+                    ['role_id = ?' => $data['role_id']]
                 );
             }
         }
@@ -313,7 +314,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             'sort_order' => 0,
             'role_type'  => Mage_Admin_Model_Acl::ROLE_TYPE_USER,
             'user_id'    => $user->getUserId(),
-            'role_name'  => $user->getFirstname(),
+            'role_name'  => $user->getFirstname()
         ]);
 
         $insertData = $this->_prepareDataForTable($data, $this->getTable('admin/role'));
@@ -330,6 +331,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Delete user role
      *
+     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
      * @return $this
      */
     public function deleteFromRole(Mage_Core_Model_Abstract $user)
@@ -337,6 +339,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
         if ($user->getUserId() <= 0) {
             return $this;
         }
+
         if ($user->getRoleId() <= 0) {
             return $this;
         }
@@ -355,6 +358,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Check if role user exists
      *
+     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
      * @return array
      */
     public function roleUserExists(Mage_Core_Model_Abstract $user)
@@ -382,6 +386,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Check if user exists
      *
+     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
      * @return array|false
      */
     public function userExists(Mage_Core_Model_Abstract $user)
@@ -415,7 +420,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             $this->_getWriteAdapter()->update(
                 $this->getMainTable(),
                 ['extra' => $data],
-                ['user_id = ?' => (int) $object->getId()],
+                ['user_id = ?' => (int) $object->getId()]
             );
         }
 
@@ -435,13 +440,13 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
             $this->_getWriteAdapter()->update(
                 $this->getMainTable(),
                 ['reload_acl_flag' => $flag],
-                ['user_id = ?' => (int) $object->getId()],
+                ['user_id = ?' => (int) $object->getId()]
             );
             if ($flag) {
                 // refresh cache menu
                 Mage::app()->getCache()->clean(
                     Zend_Cache::CLEANING_MODE_MATCHING_TAG,
-                    [Mage_Adminhtml_Block_Page_Menu::CACHE_TAGS],
+                    [Mage_Adminhtml_Block_Page_Menu::CACHE_TAGS]
                 );
             }
         }
@@ -452,6 +457,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
     /**
      * Unserializes user extra data
      *
+     * @param Mage_Core_Model_Abstract|Mage_Admin_Model_User $user
      * @return Mage_Core_Model_Abstract
      */
     protected function _unserializeExtraData(Mage_Core_Model_Abstract $user)
@@ -462,6 +468,7 @@ class Mage_Admin_Model_Resource_User extends Mage_Core_Model_Resource_Db_Abstrac
         } catch (Exception $e) {
             $user->setExtra(false);
         }
+
         return $user;
     }
 }
